refactor(logic): move bootstrap to its own package
...to avoir circular dependencies when testing
