+++ x/logic/interpreter/bootstrap/bootstrap.go
@@ -1,4 +1,4 @@
-package interpreter
+package bootstrap
 
 import _ "embed"
