+++ x/logic/predicate/address_test.go
@@ -117,7 +117,7 @@ func TestBech32(t *testing.T) {
 					ctx := sdk.NewContext(stateStore, tmproto.Header{}, false, log.NewNopLogger())
 
 					Convey("and a vm", func() {
-						interpreter := testutil.NewInterpreterMust(ctx)
+						interpreter := testutil.NewLightInterpreterMust(ctx)
 						interpreter.Register2(engine.NewAtom("bech32_address"), Bech32Address)
 
 						err := interpreter.Compile(ctx, tc.program)
